Fix Split Two Columns files name case
We should use camel case
